クラス実装っぽいモジュールをクラス実装にしてみる
